eQTLmapping_susie: Function to perform eQTL fine-mapping using SuSiE and optional resampling

This function performs fine-mapping of eQTLs using the SuSiE (Sum of Single Effects) algorithm. It allows for optional resampling of eQTL effects and returns both the estimated effects and resampled effects.

Usage

eQTLmapping_susie(
  bX,
  LD,
  Nvec,
  pip_thres = 0.5,
  pip.min = 0.2,
  L = 8,
  coverage = 0.95,
  max_iter = 300,
  ...
)

Arguments

bX

A matrix of Z-scores of marginal eQTL effect estimates for tissue-gene pairs.

LD

The LD matrix of variants.

Nvec

A vector representing the sample sizes of tissue-gene pair eQTL studies.

pip_thres

A threshold for individual PIP when no credible set is found. Default is 0.2.

pip.min

The minimum individual PIP in each 95% credible set. Used to remove variables with low PIPs within credible sets. Default is 0.05.

L

The number of single effects to be used in the SuSiE model. Default is 8.

coverage

The coverage of credible set to be used in SuSiE. Default is 0.95.

max_iter

The maximum iterations in the SuSiE model. Default is 300.

Value

A matrix of estimated eQTL effects for tissue-gene pairs.