poet_shrinkage: Function to perform covariance matrix shrinkage using the POET method

This function performs shrinkage of the covariance matrix using the POET (Principal Orthogonal complEment Thresholding) method, with an automatic selection of the number of latent factors and the optimal shrinkage parameter.

Usage

poet_shrinkage(
  LD,
  KMax = min(15, round(nrow(LD)/2)),
  lamvec = seq(0.025, 0.25, by = 0.025),
  minvalue = 0.001
)

Arguments

LD

A covariance matrix of linkage disequilibrium (LD).

KMax

The maximum number of latent factors used in POET. Default is `min(15, round(nrow(LD)/2))`.

lamvec

A vector of candidate shrinkage parameters. Default is `seq(0.025, 0.25, by = 0.025)`.

minvalue

The threshold for the minimum eigenvalues used in determining the optimal shrinkage parameter. Default is `1e-3`.

Value

A covariance matrix that has been shrunk using the POET method.